app/backend/requirements.txt
aerich==0.5.8
asyncpg==0.25.0
bcrypt==3.2.0
fastapi==0.70.0
passlib==1.7.4
python-dotenv==0.19.2
python-multipart==0.0.5
python-jose==3.3.0
requests==2.26.0
tortoise-orm==0.17.8
uvicorn==0.15.0
zeep==4.1.0